Basic request and reply functions created. Server is using a dummy reply in all resquests.
Class description:
Mosaic: Main class that starts the server.
HttpServer: Class which configures the connections using sockets, receive the requests and send a reply.
Request: Class that handle the information in income.
Reply: Class that handle the reply to the browser.

www folder: Contains all the files to send in case a reply.